refactor: Update all flat files to import from .protocol
Step 2 of Issue #320 Phase 1: Protocol import unification.

Changes:
- Updated 14 flat files to import from .protocol instead of .geometry_protocol
  - simple_grid.py, simple_grid_1d.py, tensor_product_grid.py
  - base.py, base_geometry.py
  - amr_1d.py, amr_triangular_2d.py, amr_quadtree_2d.py, amr_tetrahedral_3d.py
  - point_cloud.py, network_geometry.py
  - graph/maze_generator.py, mazes/maze_generator.py
  - implicit/implicit_domain.py

All geometry files now use unified protocol.py for imports.
Backward compatibility maintained via geometry_protocol.py re-exports.

Related to Issue #320
