+++ mfg_pde/geometry/amr_quadtree_2d.py
@@ -14,7 +14,7 @@ from typing import TYPE_CHECKING, Any
 
 import numpy as np
 
-from .geometry_protocol import GeometryType
+from .protocol import GeometryType
 
 # Apply TYPE_CHECKING isolation principle for JAX (same as OmegaConf pattern)
 if TYPE_CHECKING:
