+++ mfg_pde/geometry/amr_tetrahedral_3d.py
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
     from numpy.typing import NDArray
 
 from .base_geometry import MeshData
-from .geometry_protocol import GeometryType
+from .protocol import GeometryType
 
 
 @dataclass
