+++ mfg_pde/geometry/base_geometry.py
@@ -14,7 +14,7 @@ from typing import TYPE_CHECKING, Any
 
 import numpy as np
 
-from .geometry_protocol import GeometryType
+from .protocol import GeometryType
 
 if TYPE_CHECKING:
     from numpy.typing import NDArray
