+++ mfg_pde/geometry/implicit/implicit_domain.py
@@ -24,7 +24,7 @@ import numpy as np
 from numpy.typing import NDArray
 
 from mfg_pde.geometry.base import ImplicitGeometry
-from mfg_pde.geometry.geometry_protocol import GeometryType
+from mfg_pde.geometry.protocol import GeometryType
 
 
 class ImplicitDomain(ImplicitGeometry):
