+++ mfg_pde/geometry/tensor_product_grid.py
@@ -24,7 +24,7 @@ from typing import TYPE_CHECKING
 import numpy as np
 
 from .base import CartesianGrid
-from .geometry_protocol import GeometryType
+from .protocol import GeometryType
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Sequence
